Remove plugin interface from test_drawing
This was never used and doesn't seem like it ever will be.
