+++ tests/drawing/drawing.h
@@ -9,23 +9,44 @@
 #ifndef _WX_TESTS_DRAWING_H_
 #define _WX_TESTS_DRAWING_H_
 
-#include "gcfactory.h"
+// wxCairoRenderer::CreateMeasuringContext() is not implemented for wxX11
+#if wxUSE_GRAPHICS_CONTEXT && !defined(__WXX11__)
+    #define wxUSE_TEST_GC_DRAWING 1
+#else
+    #define wxUSE_TEST_GC_DRAWING 0
+#endif
 
 #if wxUSE_TEST_GC_DRAWING
 
+#include "wx/filename.h"
+#include "wx/graphics.h"
 #include "wx/math.h"
 
 #if wxUSE_SVG
 #include "wx/dcsvg.h"
 #endif
 
-class WXDLLIMPEXP_FWD_BASE wxDynamicLibrary;
+// Implement a class derived from this one to let the drawing test create
+// graphics context, save them and destroy them.
+class DrawingTestGCFactory
+{
+public:
+    virtual ~DrawingTestGCFactory() {}
+
+    virtual wxString GetIdForFileName () const = 0;
+    virtual wxString GetExtensionForFileName () const = 0;
+    virtual bool UseImageComparison() const = 0;
+    virtual bool PlatformIndependent() const = 0;
+    virtual wxGraphicsContext *BuildNewContext (wxSize expectedSize,
+        double pointsPerInch, const wxFileName &targetFileName) = 0;
+    virtual void SaveBuiltContext (wxGraphicsContext *&gc) = 0;
+    virtual void CleanUp (wxGraphicsContext *gc) = 0;
+};
 
 class GraphicsContextDrawingTestCase : public CppUnit::TestCase
 {
 public:
     GraphicsContextDrawingTestCase()
-        : m_drawingPluginsLoaded(false)
     {
         wxImage::AddHandler (new wxPNGHandler());
         wxImage::AddHandler (new wxBMPHandler());
@@ -44,8 +65,6 @@ private:
 #if wxUSE_SVG
 //        CPPUNIT_TEST( DrawToSVG_Basics );
 #endif
-        CPPUNIT_TEST( DrawToPlugins_Basics );
-
         // FIXME: Reference data files are currently not found when using Unix
         // build system, so these tests are failing there, fix this and remove
         // this ifdef.
@@ -54,7 +73,6 @@ private:
 #if wxUSE_SVG
 //        CPPUNIT_TEST( DrawToSVG_Fonts );
 #endif
-        CPPUNIT_TEST( DrawToPlugins_Fonts );
 #endif // __WINDOWS__
    CPPUNIT_TEST_SUITE_END();
 
@@ -122,15 +140,6 @@ private:
         DrawingTestGCFactory& gcFactory,
         const DrawingTestCase & testCase);
 
-    // enumerates the dll names as specified in WX_TEST_SUITE_GC_DRAWING_PLUGINS
-    //  (coma separated list of DLL to load and test)
-    // each DLL should have these procedures:
-    //  - DrawingTestGCFactory* CreateDrawingTestLifeCycle(),
-    //          to create the life cycle object
-    //  - DestroyDrawingTestLifeCycle(DrawingTestGCFactory *),
-    //          to destroy the life cycle object
-    void RunPluginsDrawingCase (const DrawingTestCase & testCase);
-
     void DrawToImage_Basics() {
         RunIndividualDrawingCase (ms_imageLifeCycle, ms_drawingBasicTc);
     }
@@ -139,9 +148,6 @@ private:
         RunIndividualDrawingCase (ms_svgLifeCycle, ms_drawingBasicTc);
     }
 #endif
-    void DrawToPlugins_Basics() {
-        RunPluginsDrawingCase (ms_drawingBasicTc);
-    }
     void DrawToImage_Fonts() {
         RunIndividualDrawingCase (ms_imageLifeCycle, ms_drawingFontTc);
     }
@@ -150,9 +156,6 @@ private:
         RunIndividualDrawingCase (ms_svgLifeCycle, ms_drawingFontTc);
     }
 #endif
-    void DrawToPlugins_Fonts() {
-        RunPluginsDrawingCase (ms_drawingFontTc);
-    }
 
     bool GetBuildReference() const;
 
@@ -162,19 +165,6 @@ private:
 
     static wxString ms_referenceDirectory;
     static bool ms_buildReference, ms_buildReferenceDetermined;
-    bool m_drawingPluginsLoaded;
-
-    struct PluginInfo {
-        PluginInfo();
-        ~PluginInfo();
-
-        wxDynamicLibrary* library;
-
-        DrawingTestGCFactory *gcFactory;
-        DestroyDrawingTestLifeCycleFunction destroy;
-    };
-
-    wxVector<PluginInfo> m_drawingPlugins;
 
     wxDECLARE_NO_COPY_CLASS(GraphicsContextDrawingTestCase);
 };
