+++ tests/makefile.vc
@@ -135,22 +135,10 @@ TEST_DRAWING_OBJECTS =  \
 	$(OBJS)\test_drawing_dummy.obj \
 	$(OBJS)\test_drawing_test.obj \
 	$(OBJS)\test_drawing_drawing.obj \
-	$(OBJS)\test_drawing_plugindriver.obj \
 	$(OBJS)\test_drawing_basictest.obj \
 	$(OBJS)\test_drawing_fonttest.obj
 TEST_DRAWING_RESOURCES =  \
 	$(OBJS)\test_drawing_test.res
-TEST_DRAWINGPLUGIN_CXXFLAGS = /M$(__RUNTIME_LIBS_45)$(__DEBUGRUNTIME) /DWIN32 \
-	$(__DEBUGINFO) /Fd$(OBJS)\test_drawingplugin.pdb $(____DEBUGRUNTIME) \
-	$(__OPTIMIZEFLAG) /D_CRT_SECURE_NO_DEPRECATE=1 \
-	/D_CRT_NON_CONFORMING_SWPRINTFS=1 /D_SCL_SECURE_NO_WARNINGS=1 \
-	$(__NO_VC_CRTDBG_p) $(__TARGET_CPU_COMPFLAG_p) /D__WXMSW__ \
-	$(__WXUNIV_DEFINE_p) $(__DEBUG_DEFINE_p) $(__NDEBUG_DEFINE_p) \
-	$(__EXCEPTIONS_DEFINE_p) $(__RTTI_DEFINE_p) $(__THREAD_DEFINE_p) \
-	/I$(SETUPHDIR) /I.\..\include $(____CAIRO_INCLUDEDIR_FILENAMES) /W4 \
-	$(__DLLFLAG_p) $(__RTTIFLAG) $(__EXCEPTIONSFLAG) $(CPPFLAGS) $(CXXFLAGS)
-TEST_DRAWINGPLUGIN_OBJECTS =  \
-	$(OBJS)\test_drawingplugin_pluginsample.obj
 TEST_GUI_CXXFLAGS = /M$(__RUNTIME_LIBS_59)$(__DEBUGRUNTIME) /DWIN32 \
 	$(__DEBUGINFO) /Fd$(OBJS)\test_gui.pdb $(____DEBUGRUNTIME) \
 	$(__OPTIMIZEFLAG) /D_CRT_SECURE_NO_DEPRECATE=1 \
@@ -489,9 +477,6 @@ __RUNTIME_LIBS_27 = $(__THREADSFLAG)
 !if "$(MONOLITHIC)" == "1" && "$(USE_STC)" == "1"
 __LIB_LEXILLA_IF_MONO_p = $(__LIB_LEXILLA_p)
 !endif
-!if "$(SHARED)" == "1" && "$(USE_GUI)" == "1"
-__test_drawingplugin___depname = $(OBJS)\test_drawingplugin.dll
-!endif
 !if "$(BUILD)" == "debug" && "$(DEBUG_INFO)" == "default"
 __DEBUGINFO_37 = $(__DEBUGRUNTIME_3)
 !endif
@@ -828,7 +813,7 @@ $(OBJS):
 
 ### Targets: ###
 
-all: $(OBJS)\test.exe $(__test_drawing___depname) $(__test_drawingplugin___depname) $(__test_gui___depname) $(__test_allheaders___depname) data data-image-sample data-images en_GB fr ja xart-dothraki
+all: $(OBJS)\test.exe $(__test_drawing___depname) $(__test_gui___depname) $(__test_allheaders___depname) data data-image-sample data-images en_GB fr ja xart-dothraki
 
 clean: 
 	-if exist $(OBJS)\*.obj del $(OBJS)\*.obj
@@ -840,9 +825,6 @@ clean:
 	-if exist $(OBJS)\test_drawing.exe del $(OBJS)\test_drawing.exe
 	-if exist $(OBJS)\test_drawing.ilk del $(OBJS)\test_drawing.ilk
 	-if exist $(OBJS)\test_drawing.pdb del $(OBJS)\test_drawing.pdb
-	-if exist $(OBJS)\test_drawingplugin.dll del $(OBJS)\test_drawingplugin.dll
-	-if exist $(OBJS)\test_drawingplugin.ilk del $(OBJS)\test_drawingplugin.ilk
-	-if exist $(OBJS)\test_drawingplugin.pdb del $(OBJS)\test_drawingplugin.pdb
 	-if exist $(OBJS)\test_gui.exe del $(OBJS)\test_gui.exe
 	-if exist $(OBJS)\test_gui.ilk del $(OBJS)\test_gui.ilk
 	-if exist $(OBJS)\test_gui.pdb del $(OBJS)\test_gui.pdb
@@ -859,13 +841,6 @@ $(OBJS)\test.exe: $(OBJS)\test_dummy.obj  $(TEST_OBJECTS) $(OBJS)\test_test.res
 $(OBJS)\test_drawing.exe: $(OBJS)\test_drawing_dummy.obj  $(TEST_DRAWING_OBJECTS) $(OBJS)\test_drawing_test.res
 	link /NOLOGO /OUT:$@  $(__DEBUGINFO_5) /pdb:"$(OBJS)\test_drawing.pdb" $(__DEBUGINFO_19)  $(LINK_TARGET_CPU) /LIBPATH:$(LIBDIRNAME) /SUBSYSTEM:CONSOLE    $(____CAIRO_LIBDIR_FILENAMES) $(LDFLAGS) @<<
 	$(TEST_DRAWING_OBJECTS) $(TEST_DRAWING_RESOURCES)  $(__WXLIB_CORE_p)  $(__WXLIB_NET_p)  $(__WXLIB_BASE_p)  $(__WXLIB_MONO_p) $(__LIB_SCINTILLA_IF_MONO_p) $(__LIB_LEXILLA_IF_MONO_p) $(__LIB_TIFF_p) $(__LIB_JPEG_p) $(__LIB_PNG_p) $(__LIB_WEBP_p) $(__LIB_LUNASVG_p)   wxzlib$(WXDEBUGFLAG).lib wxregexu$(WXDEBUGFLAG).lib wxexpat$(WXDEBUGFLAG).lib $(EXTRALIBS_FOR_BASE) $(__CAIRO_LIB_p) kernel32.lib user32.lib gdi32.lib gdiplus.lib msimg32.lib comdlg32.lib winspool.lib winmm.lib shell32.lib shlwapi.lib comctl32.lib ole32.lib oleaut32.lib uuid.lib rpcrt4.lib advapi32.lib version.lib ws2_32.lib wininet.lib
-<<
-!endif
-
-!if "$(SHARED)" == "1" && "$(USE_GUI)" == "1"
-$(OBJS)\test_drawingplugin.dll: $(TEST_DRAWINGPLUGIN_OBJECTS)
-	link /DLL /NOLOGO /OUT:$@  $(__DEBUGINFO_5) /pdb:"$(OBJS)\test_drawingplugin.pdb" $(__DEBUGINFO_37)  $(LINK_TARGET_CPU) /LIBPATH:$(LIBDIRNAME)   $(____CAIRO_LIBDIR_FILENAMES) $(LDFLAGS) @<<
-	$(TEST_DRAWINGPLUGIN_OBJECTS)   $(__WXLIB_CORE_p)  $(__WXLIB_BASE_p)  $(__WXLIB_MONO_p) $(__LIB_SCINTILLA_IF_MONO_p) $(__LIB_LEXILLA_IF_MONO_p_1) $(__LIB_TIFF_p) $(__LIB_JPEG_p) $(__LIB_PNG_p) $(__LIB_WEBP_p) $(__LIB_LUNASVG_p)   wxzlib$(WXDEBUGFLAG).lib wxregexu$(WXDEBUGFLAG).lib wxexpat$(WXDEBUGFLAG).lib $(EXTRALIBS_FOR_BASE) $(__CAIRO_LIB_p) kernel32.lib user32.lib gdi32.lib gdiplus.lib msimg32.lib comdlg32.lib winspool.lib winmm.lib shell32.lib shlwapi.lib comctl32.lib ole32.lib oleaut32.lib uuid.lib rpcrt4.lib advapi32.lib version.lib ws2_32.lib wininet.lib
 <<
 !endif
 
@@ -1187,17 +1162,11 @@ $(OBJS)\test_drawing_test.obj: .\test.cpp
 $(OBJS)\test_drawing_drawing.obj: .\drawing\drawing.cpp
 	$(CXX) /c /nologo /TP /Fo$@ $(TEST_DRAWING_CXXFLAGS) .\drawing\drawing.cpp
 
-$(OBJS)\test_drawing_plugindriver.obj: .\drawing\plugindriver.cpp
-	$(CXX) /c /nologo /TP /Fo$@ $(TEST_DRAWING_CXXFLAGS) .\drawing\plugindriver.cpp
-
 $(OBJS)\test_drawing_basictest.obj: .\drawing\basictest.cpp
 	$(CXX) /c /nologo /TP /Fo$@ $(TEST_DRAWING_CXXFLAGS) .\drawing\basictest.cpp
 
 $(OBJS)\test_drawing_fonttest.obj: .\drawing\fonttest.cpp
 	$(CXX) /c /nologo /TP /Fo$@ $(TEST_DRAWING_CXXFLAGS) .\drawing\fonttest.cpp
-
-$(OBJS)\test_drawingplugin_pluginsample.obj: .\drawing\pluginsample.cpp
-	$(CXX) /c /nologo /TP /Fo$@ $(TEST_DRAWINGPLUGIN_CXXFLAGS) .\drawing\pluginsample.cpp
 
 $(OBJS)\test_gui_dummy.obj: .\dummy.cpp
 	$(CXX) /c /nologo /TP /Fo$@ $(TEST_GUI_CXXFLAGS) /Yctestprec.h .\dummy.cpp
